requirements.txt
dvc
dvc[gdrive]
tensorflow-cpu==2.2.0
pandas
pytest
tox
flake8
Flask
gunicorn
mlflow